backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4

alembic==1.13.2

httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.9
